+++ src/dataobjects/mysterydata.cc
@@ -98,7 +98,7 @@ void MysteryData::dump_out(FILE *f,int indent,int what,Laxkit::DumpContext *cont
 	fprintf(f,"%smaxx %.10g\n",spc,maxx);
 	fprintf(f,"%smaxy %.10g\n",spc,maxy);
 	fprintf(f,"%smatrix %.10g %.10g %.10g %.10g %.10g %.10g\n",spc,
-				m(0),m(1),m(2),m(3),m(4),m(5));
+				m((int)0),m(1),m(2),m(3),m(4),m(5));
 	fprintf(f,"%snativeid %ld\n",spc,nativeid);
 	if (attributes) {
 		fprintf(f,"%sattributes\n",spc);
@@ -127,7 +127,7 @@ Laxkit::Attribute *MysteryData::dump_out_atts(Laxkit::Attribute *att,int what,La
 	att->push("maxx", maxx);
 	att->push("maxy", maxy);
 	att->pushStr("matrix", -1, "%.10g %.10g %.10g %.10g %.10g %.10g",
-				m(0),m(1),m(2),m(3),m(4),m(5));
+				m((int)0),m(1),m(2),m(3),m(4),m(5));
 	att->push("nativeid", nativeid);
 	if (attributes) {
 		Attribute *att2 = attributes->duplicate();
