+++ src/polyptych/src/nets.cc
@@ -1235,7 +1235,7 @@ void Net::dump_out(FILE *f,int indent,int what,Laxkit::DumpContext *context)
 	fprintf(f,"%sinfo %d\n",spc,info);
 	
 	fprintf(f,"%smatrix %.10g %.10g %.10g %.10g %.10g %.10g\n",
-		            spc,m(0),m(1),m(2),m(3),m(4),m(5));
+		            spc,m((int)0),m(1),m(2),m(3),m(4),m(5));
 
 
 	if (basenet) {
